requirements.txt
streamlit
mistralai==0.4.2
python-dotenv